master

+++ src/components/Header/OrderItem.js
@@ -1,16 +1,30 @@
 import React from 'react';
-import {useDispatch} from "react-redux";
+import {useDispatch, useSelector} from "react-redux";
 import {ImCross} from "react-icons/im";
+import {CiCircleMinus, CiCirclePlus} from "react-icons/ci";
 
-import {deleteItemFromCart} from "../../redux";
+
+import {deleteItemFromCart, minusItem, setItemInCart} from "../../redux";
 
 import css from '..//../pages/OrderPage/OrderPage.module.css'
 
 const OrderItem = ({product}) => {
+    const {id} = product;
+
+    const {itemsInCart, quantity} = useSelector(state => state.cart);
 
     const dispatch = useDispatch();
-    const hadleClick = () => {
-        dispatch(deleteItemFromCart(product.id))
+
+    const onClickPlus = () => {
+        dispatch(setItemInCart({id}))
+    }
+
+    const onClickMinus = () => {
+        dispatch(minusItem(id))
+    }
+    const onClickRemove = () => {
+        if (window.confirm('Are you sure you want to remove ?'))
+            dispatch(deleteItemFromCart(product.id))
     }
 
     return (
@@ -21,12 +35,24 @@ const OrderItem = ({product}) => {
             <div className={css.OrderItemTitle}>
                 <span>{product.title}</span>
             </div>
+            <div className={css.TotalAddingProducts}>
+                <CiCirclePlus
+                    size={22}
+                    className={css.CartItemPlusIcon}
+                    onClick={onClickPlus}
+                /> {product.count}
+                <CiCircleMinus
+                    size={22}
+                    className={css.CartItemMinusIcon}
+                    onClick={onClickMinus}
+                />
+            </div>
             <div className={css.OrderItemPrice}>
-                <span>{product.price} $</span>
+                <span>{product.price * product.count} $</span>
                 <ImCross
-                    size={20}
+                    size={17}
                     className={css.CartItemDeleteIcon}
-                    onClick={hadleClick}
+                    onClick={onClickRemove}
                 />
             </div>
         </div>
